pyproject.toml
[project]
name = "revin-techno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.20",
    "fastapi>=0.116.1",
    "gtts>=2.5.4",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.10",
    "langchain-openai>=0.3.32",
    "openai>=1.102.0",
    "pdfplumber>=0.11.7",
    "pyaudio>=0.2.14",
    "pydantic>=2.11.7",
    "pydub>=0.25.1",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "speechrecognition>=3.14.3",
    "streamlit>=1.49.1",
    "uvicorn[standard]>=0.35.0",
]
